+++ midilib/src/main/cpp/oscillators/Oscillator.h
@@ -7,8 +7,8 @@
 
 class Oscillator {
 public:
-    virtual float calculate(float time, float frequency, int8_t overtoneFactor) = 0;
-    float eval(float time, float frequency);
+    virtual float calculate(double time, float frequency, int8_t overtoneFactor) = 0;
+    float eval(double time, float frequency);
 
     Oscillator(float amplitude, float phase, float freqFactor, int8_t overtones);
 
@@ -18,7 +18,7 @@ protected:
     float mFreqFactor = 1;
     int8_t mOvertones = 0;
 
-    float calcPhase(float time, float frequency, int8_t overtoneFactor) const;
+    float calcPhase(double time, float frequency, int8_t overtoneFactor) const;
 };
