+++ midilib/src/main/cpp/oscillators/ReverseSawtoothOscillator.h
@@ -8,7 +8,7 @@
 class ReverseSawtoothOscillator: public Oscillator {
     using Oscillator::Oscillator;
 
-    float calculate(float time, float frequency, int8_t overtoneFactor) override;
+    float calculate(double time, float frequency, int8_t overtoneFactor) override;
 };
