+++ midilib/src/main/cpp/oscillators/SinOscillator.h
@@ -8,7 +8,7 @@
 class SinOscillator: public Oscillator {
     using Oscillator::Oscillator;
 
-    float calculate(float time, float frequency, int8_t overtoneFactor) override;
+    float calculate(double time, float frequency, int8_t overtoneFactor) override;
 };
